+++ spec/Formatter/FullHttpMessageFormatterSpec.php
@@ -227,4 +227,25 @@ X-Param-Bar: bar
 STR;
         $this->formatResponse($response)->shouldReturn($expectedMessage);
     }
+
+    function it_omits_body_with_null_bytes(RequestInterface $request, StreamInterface $stream)
+    {
+        $this->beConstructedWith(1);
+
+        $stream->isSeekable()->willReturn(true);
+        $stream->rewind()->shouldBeCalled();
+        $stream->__toString()->willReturn("\0");
+        $request->getBody()->willReturn($stream);
+        $request->getMethod()->willReturn('GET');
+        $request->getRequestTarget()->willReturn('/foo');
+        $request->getProtocolVersion()->willReturn('1.1');
+        $request->getHeaders()->willReturn([]);
+
+        $expectedMessage = <<<STR
+GET /foo HTTP/1.1
+
+[binary stream omitted]
+STR;
+        $this->formatRequest($request)->shouldReturn($expectedMessage);
+    }
 }