+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+- Omitted binary body in FullHttpMessageFormatter. `[binary stream omitted]` will be shown instead.
+
 ### Added
 
 - New Header authentication method for arbitrary header authentication.